+++ src/contracts/examples/v4-example/UniV4Hook.sol
@@ -75,13 +75,13 @@ contract UniV4Hook is V4DAppControl {
 
         Context memory ctx = ISafetyLocks(ATLAS).getLockState();
 
-        if (ctx.phase == ExecutionPhase.UserOperation) {
+        if (ctx.phase == uint8(ExecutionPhase.UserOperation)) {
             // Case: User call
             // Sender = ExecutionEnvironment
 
             // Verify that the pool is valid for the user to trade in.
             require(keccak256(abi.encode(key, sender)) == hashLock, "ERR-H02 InvalidSwapper");
-        } else if (ctx.phase == ExecutionPhase.SolverOperations) {
+        } else if (ctx.phase == uint8(ExecutionPhase.SolverOperations)) {
             // Case: Solver call
             // Sender = Solver contract
             // NOTE: This phase verifies that the user's transaction has already
