+++ src/contracts/examples/v4-example/V4DAppControl.sol
@@ -210,7 +210,7 @@ contract V4DAppControl is DAppControl {
         // and that DAppControl supplied a valid signature
         require(address(this) == hook, "ERR-H00 InvalidCallee");
         require(hook == _control(), "ERR-H01 InvalidCaller");
-        require(ctx.phase == ExecutionPhase.PreOps, "ERR-H02 InvalidLockStage");
+        require(ctx.phase == uint8(ExecutionPhase.PreOps), "ERR-H02 InvalidLockStage");
         require(hashLock == bytes32(0), "ERR-H03 AlreadyActive");
 
         // Set the storage lock to block reentry / concurrent trading
@@ -228,7 +228,7 @@ contract V4DAppControl is DAppControl {
         // and that DAppControl supplied a valid signature
         require(address(this) == hook, "ERR-H20 InvalidCallee");
         require(hook == _control(), "ERR-H21 InvalidCaller");
-        require(ctx.phase == ExecutionPhase.PostOps, "ERR-H22 InvalidLockStage");
+        require(ctx.phase == uint8(ExecutionPhase.PostOps), "ERR-H22 InvalidLockStage");
         require(hashLock == keccak256(abi.encode(key, msg.sender)), "ERR-H23 InvalidKey");
 
         // Release the storage lock
